run-api-server/sites/serializers.py
from .models import (Sites, SiteSettings)
from rest_framework import serializers


class SitesSerializer(serializers.ModelSerializer):
    class Meta:
        model = Sites
        fields = '__all__'


class SiteSettingsSerializer(serializers.ModelSerializer):
    class Meta:
        model = SiteSettings
        fields = '__all__'
