run-api-server/requirements.txt
ansible==4.6.0
ansible-core==2.11.5
appdirs==1.4.4
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
boto3==1.18.35
botocore==1.21.35
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.8
decorator==4.4.2
distlib==0.3.1
Django==3.2.6
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
filelock==3.0.12
idna==2.10
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mysql-client==0.0.1
mysqlclient==2.0.1
networkx==2.5.1
packaging==21.0
platformdirs==2.2.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-http-client==3.3.2
python-jsonrpc-server==0.3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
resolvelib==0.5.4
rstcheck==3.3.1
s3transfer==0.5.0
sendgrid==6.8.1
six==1.15.0
snooty-lextudio==1.9.2.dev0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
toml==0.10.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.5
virtualenv==20.7.2
watchdog==1.0.2
zohocrmsdk2-0==4.0.0
